refactor: split image sandbox adapter
